requirements.txt
Pillow>=9.0.0
slack_sdk>=3.19.0
requests>=2.28.0 